Extend the interface to return closures returning randoms.
In order to gain about a 1.6x speedup in normal-distributed random
numbers we must avoid creating the "distribution" object on each call.
To do that, for each "direct" <distribution>(params) method that
IRandom had before it now has a make_<distribution>(params) method.
These return a function object which, when called, produces a random
like the original call would have.  However, the implementation may
provide this callable as a closure around the underlying distribution
object and thus avoid a construction/destruction for each call.