Fix #202 (again)
This revert the use of round() in Digitizer but adds a configuration option of the same name to allow the user to reenable, if they desire.  See #202 for examples of the implications of either choice.

The bug that orginally launched #202 is still fixed but, for now, AddNoise gains a configuration parameter to allow the user to activate this bug and at any desired level of damage.  Of course, default is no bug.  This is added for testing purposes and may be removed in the near future.

Change the "roundtrip" test to utilize all 10 levels of noise, each level across a contigous group of 256 channels.

The bats test is greatly expanded.  It currently requires the "bug202" option and an updated "wirecell-plot comp1d" which is to be committed shortly.