Do not store white-noise constant as spectral element.
In past, a white-noise constant specified by some spectrum files was
appended to the (half) spectrum amplitude array.  This "leaked" out
and led to some bugs while hiding others.  This internal value is now
handled explicitly and separately from the "amp" spectrum.

The spectra for each channel is cached to avoid returning refernce to
a single static value.  This caching may lead to too much memory.  The
real solution is to change the Interface to return by copy.