Fix #211.
The "measure" is found by applying the connected components algorithm
to a special per-slice, per plane graph holding nodes representing
blobs and channels.  The nodes of this graph hold the vertex
descriptor of the original cluster graph node.

The bug was due to a channel being added to this special graph each
time it was seen across the blobs.  Thus, by construction, the
connected components (ie, measures) always had exactly one blob.

The fix was to only add any given channel exactly once.

The following test is extended to check this bug:

```
$ bats img/test/test-wct-uboone-img.bats
test-wct-uboone-img.bats
 ✓ create graph
 ✓ check log files
 ✓ inspect blobs
 ✓ dump blobs
 ✓ at least one multi-blob measure

5 tests, 0 failures
```