Breaking change: include graph vertex and edge descriptors in cluster arrays.
To have any hope of developing a perfect round trip through the
cluster arrays schema the original cluster graph vertex and edge
descriptors must be included in some manner.  This will allow an
ICluster graph to be reconstructed with original vertex and edge
ordering.

This change inserts a new column zero in the node arrays to hold their
vertex descriptor.