Remove as much mention of the now eradicated "threshold" tag.
This uncovers the problem that wclsFrameSaver in larwirecell uses the
"threshold" tag to both loate the wiener RMS summary and to save it
into the art::Event.  For now I have left this, until understanding
how best to handle it.  If we also change this to "wiener" it has
downstream consequences on any consumer of the art::Event.