Last eradication of use of the "threshold" tagged trace summary in favor of "wiener".
WARNING to art/larsoft users:

The wclsFrameSaver uses this tag both to find traces and to save a
vector<double> data product in the art::Event.  This commit changes
how wclsFrameSaver is configured.  There are to implications:

1) You may have your own configuration of wclsFrameSaver.  It MUST use
"wiener" to find Wiener RMS summary as "threshold" does no longer
exist.

2) Any art/larsoft code downstream must look for these values in
"wiener" and not "threshold".

If all goes well, any old configuration will cause art/larsoft to
exist quickly and loudly so it should be easy to find the
configuration that needs updating.