Develop a new blob set reframer.
This is intended to be used as part of a test for tiling.  The tenio
test runs it and we get something out, but more work is still needed.

Note, there is also a much older BlobReframer, (which really takes
ICluster).  It is broken in the case that the blobs have lost their
connection with their original frame, such as when they are read from
file.