Apply floor truncation on Digitizer output by default.
Previously, Digizer output fractional ADC values on the assumption
that truncation was performed by subsequent file sink.  As we now
start to hook up NF+SP directly to sim, this truncation was being
omitted leading to small discrepencies.  Thanks @weihythu and Ewerton
for finding this!

The "round" option still supported with same default being false.

This commit also does some temporary fixes some of the tests and
leaves the tenio test failing.  This is due to ongoing work in the
apply-pointcloud branch.  It requires recent wire-cell-python to be
installed.